requirements.txt
python-dotenv
langchain

# OpenAI APIを利用する場合には、コメントアクティベート
# openai
# langchain-openai

# Gemini APIを利用する場合には、コメントアクティベート
google-genai
langchain-google-genai

# Ollamaを利用する場合には、コメントアクティベート
# ollama
# langchain-ollama